Use jQuery because you're checking for window.jQuery
Some sites put jQuery into noConflict mode and don't bind to $. One example is http://shop.timberland.com/ so the pinry bookmarklet would not work. It will throw a 'document does not have a method ready' error. This change just wraps all of the jQuery specific code to use the current jQuery as $.

http://api.jquery.com/jQuery.noConflict/